Moving loss function to a new module.
